src/backend/requirements.txt
Flask == 2.2.0
Flask-Cors == 3.0.10
pymongo==3.10.1
six==1.14.0
Flask-Bcrypt==0.7.1
PyJWT==1.5.3
Werkzeug==2.2.0
MarkupSafe==1.1.1
Jinja2==2.2.0
itsdangerous==1.1.0